unique var name change, Closes issues #8 and #10
The unique values check for the adaptation sampling stage has had it's
variable name changed from unique_check to n_unique.

There is now the ability to pass n_unique as an argument to the
run_stage function, which will be used to initialise a function local
variable, and then passed on to the check_adapted function (which tests
the number of unique samples for all participants)

Additionally the result of a successful check for unique values,
followed by an unsuccessful check for the ability to create an efficient
proposal distribution is no now longer to double the number of unique
values to search for, instead the procedure is to add the original
value to the current value. Ths, given an initial value of 10 and
repeated situations where the uniqueness check succeeds but the proposal
distribution creation fails we move from a situation where we would get
successive value of the `n_unique` from `[10, 20, 40, 80, ...]` which
could be bad, to instead values matching `[10, 20, 30, 40, ...]`.