New function to "augment" a sampler object
Closes #68.

When the new version  (> 0.2.1) tries to run a sampling stage on an
older sampler object (created pre-0.2.1) then there is an error due to
missing internal array in the sampler object. This new function adds the
internal array, filled with NA values, to enable the old sampler object
to continue fitting with the new version.